squelch warning

+++ c10/c10/cpu/op/CPUAll_test.cpp
@@ -61,6 +61,6 @@ TEST(CPUAll_copy, int32) {
   std::initializer_list<int32_t> data = {0, 1, 2, 3};
   Tensor x = tensor(ArrayRef<int32_t>{data}, {2, 2});
   Tensor y = empty({2,2}, int32);
-  cpu::op::copy_(y, int32, data.begin(), data.size() * int32.itemsize());
+  cpu::op::copy_(y, int32, data.begin(), static_cast<int64_t>(data.size()) * int32.itemsize());
   ASSERT_TRUE(x.equal(y));
 }
